chore: restructure as polyglot monorepo
The JS module moves into packages/js/ and the gRPC schema into a top-level
proto/ directory so it can be shared with the upcoming Java module without
duplicating the contract.

- packages/js/ — JS source, tests, build config, README (no behavior change;
  package.json main now resolves via ./src/index.js)
- proto/conditionally_execute.proto — shared gRPC schema (moved from
  plugins/proto/), referenced from packages/js/src/plugins/grpc-consensus.js
- package.json (root) — npm workspaces declaration pointing at packages/js
- .github/workflows/nodejs.yml — runs from packages/js/ working dir, with
  cache-dependency-path scoped to the workspace lockfile
- README.md (root) — monorepo overview; original README preserved at
  packages/js/README.md

No npm/runtime behavior change for consumers — the published package still
imports as 'conditionally-execute' and exports the same surface.
