+++ .github/workflows/nodejs.yml
@@ -6,6 +6,10 @@ on:
   pull_request:
     branches: [ master, test ]
 
+defaults:
+  run:
+    working-directory: packages/js
+
 jobs:
   test:
     name: Node ${{ matrix.node-version }}
@@ -25,6 +29,7 @@ jobs:
         with:
           node-version: ${{ matrix.node-version }}
           cache: 'npm'
+          cache-dependency-path: packages/js/package-lock.json
 
       - name: Install dependencies
         run: npm ci
