+++ packages/js/.prettierrc
