+++ packages/js/eslint.config.js
