+++ packages/js/src/plugins/collect-errors.js
