+++ packages/js/src/plugins/multi-threaded.js
